requirements.txt
streamlit==1.35.0
faiss-cpu==1.7.4
pdfplumber==0.10.3
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==2.6.1
